+++ controllers/courses.ts
@@ -54,15 +54,22 @@ exports.getCourse = asyncHandler(async (req:Request, res:Response, next: any) =>
 // @route POST /api/v1/bootcamps/:bootcampId/courses
 // @access Private
 
-exports.addCourse = asyncHandler(async (req:Request, res:Response, next: any) => {
+exports.addCourse = asyncHandler(async (req:any, res:Response, next: any) => {
     
     req.body.bootcamp = req.params.bootcampId
+    req.body.user = req.user.id;
 
     const bootcamp:BootcampDocumentInterface|null = await Bootcamp.findById(req.params.bootcampId)
 
     if(!bootcamp){
         return next(new ErrorResponse(`No bootcamp with the id ${req.params.id}`),404)
     }
+
+    //make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !=='admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add course to this bootcamp`,403))
+    }
+
     const course:CourseDocInterface = await Course.create(req.body)
     res.status(200).json({
         success: true,
@@ -75,12 +82,17 @@ exports.addCourse = asyncHandler(async (req:Request, res:Response, next: any) =>
 // @route PUT /api/v1/courses/:id
 // @access Private
 
-exports.updateCourse = asyncHandler(async (req:Request, res:Response, next: any) => {
+exports.updateCourse = asyncHandler(async (req:any, res:Response, next: any) => {
 
     let course:CourseDocInterface|null = await Course.findById(req.params.id)
 
     if(!course){
         return next(new ErrorResponse(`No course with the id ${req.params.id}`),404)
+    }
+
+    //make sure user is bootcamp owner
+    if(course.user.toString() !== req.user.id && req.user.role !=='admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update course ${course._id}`,403))
     }
     
     course = await Course.findByIdAndUpdate(req.params.id, req.body,{
@@ -98,12 +110,16 @@ exports.updateCourse = asyncHandler(async (req:Request, res:Response, next: any)
 // @route DELETE /api/v1/courses/:id
 // @access Private
 
-exports.deleteCourse = asyncHandler(async (req:Request, res:Response, next: any) => {
+exports.deleteCourse = asyncHandler(async (req:any, res:Response, next: any) => {
 
     const course:CourseDocInterface|null = await Course.findById(req.params.id)
 
     if(!course){
         return next(new ErrorResponse(`No course with the id ${req.params.id}`),404)
+    }
+    //make sure user is bootcamp owner
+    if(course.user.toString() !== req.user.id && req.user.role !=='admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}`,403))
     }
     
     await course.deleteOne()
